Split main into files
