+++ Makefile
@@ -1,6 +1,8 @@
 NAME	= minishell
 
-SRC		= main.c signal.c env_list.c token_list.c lexer.c
+SRC		= main.c signal.c env_list.c token_list.c lexer.c \
+		  handle_execution.c list_to_array.c list_to_envp.c \
+		  get_pathname.c list_free.c
 
 OBJ		= $(SRC:.c=.o)
