+++ lexer.c
@@ -6,7 +6,7 @@
 /*   By: almelo <almelo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/19 13:40:12 by almelo            #+#    #+#             */
-/*   Updated: 2023/02/22 16:09:58 by almelo           ###   ########.fr       */
+/*   Updated: 2023/02/25 16:18:36 by almelo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
