+++ signal.c
@@ -6,14 +6,15 @@
 /*   By: almelo <almelo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/07 15:43:30 by almelo            #+#    #+#             */
-/*   Updated: 2023/02/22 12:27:22 by almelo           ###   ########.fr       */
+/*   Updated: 2023/02/25 16:21:32 by almelo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-static void	sigint_handler()
+static void	sigint_handler(int signo)
 {
+	(void)signo;
 	ft_putchar_fd('\n', STDOUT_FILENO);
 	rl_replace_line("", 0);
 	rl_on_new_line();
@@ -29,7 +30,6 @@ void	set_signal_handler(void)
 	sigint_act.sa_flags = SA_RESTART;
 	sigemptyset(&sigint_act.sa_mask);
 	sigaction(SIGINT, &sigint_act, NULL);
-
 	sigquit_act.sa_handler = SIG_IGN;
 	sigquit_act.sa_flags = SA_RESTART;
 	sigemptyset(&sigquit_act.sa_mask);
